Remove LCDD Geometry conversion. Keep only extraction of visualization attributes
