+++ examples/SimpleDetector/CMakeLists.txt
@@ -36,7 +36,7 @@ foreach (test Simple_ILD)
 ##    COMMAND    "${CMAKE_INSTALL_PREFIX}/bin/run_test_SimpleDetector.sh"
 ##    EXEC_ARGS  geoConverter -compact2gdml -input file:${CMAKE_CURRENT_SOURCE_DIR}/compact/${test}.xml -output file:${test}.gdml
 ##    REGEX_PASS " Successfully extracted GDML to")
-  foreach( type description vis )
+  foreach( type vis )
     dd4hep_add_test_reg( SimpleDetector_converter_${type}
       COMMAND    "${CMAKE_INSTALL_PREFIX}/bin/run_test_SimpleDetector.sh"
       EXEC_ARGS  geoConverter -compact2${type} -input file:${CMAKE_CURRENT_SOURCE_DIR}/compact/${test}.xml -output file:${test}.${type}
